Allows stripping people by resizing (#19038)
* Allows stripping people by resizing

Added a new preference that allows you to drop all of your items and possibly underwear upon being suddenly resized by more than 30%.

This also adds a variable to resize() that determines whether a specific method can trigger the stripping.
It is currently enabled on size guns and bluespace crackers only, I want to add it on a case by case basis.

* More info in tooltip

* Fix mistake

* Fix the other one too

* modern style

* Update vore_prefs.dm

* Turns defines into numbers

* Makes the new preference a dropdown

The edit mode is forced to true as the preferences page does not use edit modes.
Tested and it works fine.

* dropdown

* sublist

* .

* string keys

* tooltip

* .

* Current progress (broken)

* Fixes things

Tested and they work as expected

* please never do this

* .

* Make belly temps save

* move by 1 second

---------

+++ code/datums/ghost_spawn.dm
@@ -18,7 +18,8 @@ GLOBAL_VAR_INIT(allowed_ghost_spawns, 2)
 	if(isobserver(user))
 		var/mob/observer/dead/observer = user
 		observer.selecting_ghostrole = FALSE
-	qdel(src)
+	if(!QDELETED(src))
+		qdel(src)
 
 /datum/tgui_module/ghost_spawn_menu/tgui_interact(mob/user, datum/tgui/ui, datum/tgui/parent_ui)
 	. = ..()
